+++ CHANGELOG.md
@@ -29,6 +29,7 @@ to [Common Changelog](https://common-changelog.org)
 - Add orthogroups to genomic island cluster overview. ([#173](https://github.com/metagenlab/zDB/pull/173)) (Niklaus Johner)
 - Add genomic island prediction consistency and clustering to analysis pipeline. ([#162](https://github.com/metagenlab/zDB/pull/162)) (Niklaus Johner)
 - Add genomic island visualizations. ([#165](https://github.com/metagenlab/zDB/pull/165)) (Niklaus Johner)
+- Add script to export a loci map between genomes. ([#170](https://github.com/metagenlab/zDB/pull/170)) (Niklaus Johner)
 
 ## 1.3.5 - 2025-03-26